ml/requirements.txt
flask==3.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-decouple==3.8
gunicorn==21.2.0
python-dateutil==2.8.2
